Feat[#250] : OuterIndex
